fix: rename consul_metrics -> consul_agent (#10876)
