+++ plugins/inputs/all/all.go
@@ -28,7 +28,7 @@ import (
 	_ "github.com/influxdata/telegraf/plugins/inputs/cloudwatch"
 	_ "github.com/influxdata/telegraf/plugins/inputs/conntrack"
 	_ "github.com/influxdata/telegraf/plugins/inputs/consul"
-	_ "github.com/influxdata/telegraf/plugins/inputs/consul_metrics"
+	_ "github.com/influxdata/telegraf/plugins/inputs/consul_agent"
 	_ "github.com/influxdata/telegraf/plugins/inputs/couchbase"
 	_ "github.com/influxdata/telegraf/plugins/inputs/couchdb"
 	_ "github.com/influxdata/telegraf/plugins/inputs/cpu"
