+++ plugins/inputs/consul_agent/testdata/response_key_metrics.json
